requirements.txt
paho-mqtt==1.6.1
PySide6==6.9.0 